# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Dashboard Perumahan Lumajang
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LumajangSummary {
  totalLokasi: number;
  totalDeveloper: number;
  totalStok: number;
  totalTerjual: number;
  totalSisa: number;
  totalPeminatan: number;
  lastUpdated: string;
}

export interface KecamatanData {
  kodeWilayah: string;
  namaWilayah: string;
  supply: number;
  peminatan: number;
  pilihan: number;
  sisa: number;
}

export interface ListingItem {
  idLokasi: string;
  namaPerumahan: string;
  jenisPerumahan: string;
  kecamatan: string;
  /** @nullable */
  kelurahan: string | null;
  namaDeveloper: string;
  asosiasi: string;
  /** @nullable */
  jumlahUnit: string | null;
  foto: string[];
}

export interface DeveloperData {
  namaDeveloper: string;
  asosiasi: string;
  jumlahLokasi: number;
  totalUnit: number;
  listings: ListingItem[];
}

export interface ListingsResponse {
  data: ListingItem[];
  total: number;
  page: number;
  limit: number;
}

export interface ListingDetail {
  idLokasi: string;
  namaPerumahan: string;
  jenisPerumahan: string;
  kecamatan: string;
  /** @nullable */
  kelurahan: string | null;
  namaDeveloper: string;
  asosiasi: string;
  /** @nullable */
  jumlahUnit: string | null;
  foto: string[];
}

export interface RefreshResult {
  success: boolean;
  message: string;
  timestamp: string;
}

export type GetLumajangListingsParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
kecamatan?: string | null;
};