# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Dashboard Perumahan Lumajang
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lumajang/listings:
    get:
      operationId: getLumajangListings
      tags: [lumajang]
      summary: Daftar lokasi perumahan di Lumajang
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: kecamatan
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
  /lumajang/listings/{idLokasi}:
    get:
      operationId: getLumajangListingDetail
      tags: [lumajang]
      summary: Detail lokasi perumahan
      parameters:
        - name: idLokasi
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
...
components:
  schemas:
...
    ListingItem:
      type: object
      required:
        - idLokasi
        - namaPerumahan
        - jenisPerumahan
        - kecamatan
        - kelurahan
        - namaDeveloper
        - asosiasi
        - jumlahUnit
        - foto
      properties:
        idLokasi:
          type: string
        namaPerumahan:
          type: string
        jenisPerumahan:
          type: string
        kecamatan:
          type: string
        kelurahan:
          type: ["string", "null"]
        namaDeveloper:
          type: string
        asosiasi:
          type: string
        jumlahUnit:
          type: ["string", "null"]
        foto:
          type: array
          items:
            type: string
    ListingsResponse:
      type: object
      required:
        - data
        - total
        - page
        - limit
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ListingItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ListingDetail:
      type: object
      required:
        - idLokasi
        - namaPerumahan
        - jenisPerumahan
        - kecamatan
        - kelurahan
        - namaDeveloper
        - asosiasi
        - jumlahUnit
        - foto
      properties:
        idLokasi:
          type: string
        namaPerumahan:
          type: string
        jenisPerumahan:
          type: string
        kecamatan:
          type: string
        kelurahan:
          type: ["string", "null"]
        namaDeveloper:
          type: string
        asosiasi:
          type: string
        jumlahUnit:
          type: ["string", "null"]
        foto:
          type: array
          items:
            type: string
...