# lib/api-zod/src/generated/types/listingDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Dashboard Perumahan Lumajang
 * OpenAPI spec version: 0.1.0
 */

export interface ListingDetail {
  idLokasi: string;
  namaPerumahan: string;
  jenisPerumahan: string;
  kecamatan: string;
  /** @nullable */
  kelurahan: string | null;
  namaDeveloper: string;
  asosiasi: string;
  /** @nullable */
  jumlahUnit: string | null;
  foto: string[];
}