# lib/api-zod/src/generated/types/listingsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Dashboard Perumahan Lumajang
 * OpenAPI spec version: 0.1.0
 */
import type { ListingItem } from './listingItem';

export interface ListingsResponse {
  data: ListingItem[];
  total: number;
  page: number;
  limit: number;
}